requirements.txt
google-generativeai>=0.3.0
Pillow>=10.0.0
Flask>=3.0.0
Werkzeug>=3.0.0
reportlab>=4.0.0
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.0

